Used API to get the current weather temperatures
by using async function WeatherService the promise
is returned which is raw data. the raw data then
is processed in WeatherData class and is displayed
in the us when the city name is typed in the
search bar. the custom event is built to detect
the search and fetch the data in the AppLogic.

[Ticket: 1]